Updated the simulation files for 2D and 3D simulations
