+++ include/simulation.hpp
@@ -9,12 +9,10 @@
 #include "factories/UterineSimpleCellFactory.hpp"
 #include "factories/UterineRegularCellFactory.hpp"
 #include "factories/UterineZeroCellFactory.hpp"
-#include "factories/UterineSimpleCellFactory3d.hpp"
-#include "factories/UterineRegularCellFactory3d.hpp"
-#include "factories/UterineRegionCellFactory3d.hpp"
-#include "factories/UterineZeroCellFactory3d.hpp"
+#include "factories/UterineRegionCellFactory.hpp"
 
-void simulation_2d();
-void simulation_3d();
+void run_simulation(const int dim);
+void simulation_2d(std::string stimulus_type, std::string log_path);
+void simulation_3d(std::string stimulus_type, std::string log_path);
 
 #endif  // INCLUDE_SIMULATIONS_HPP_
